// es-ts/ts-nest-demo/src/apis/content/vo/template/activity-data-body-tag.ts
export class ActivityDataBodyTag {
  key?: string;

  skillType?: string;

  ageGroupL?: number;

  ageGroupH?: number;

  activityType?: string;

  activitySubType?: string;

  cefrLevels?: string[];

  compassTags?: string[];

  secondaryCompassTags?: string[];

  totalCompassTags?: string[];

  learningFocus?: string;
}